Potential fix for code scanning alert no. 9: Comparison of narrow type with wide type in loop condition

+++ src/cli.cpp
@@ -1059,7 +1059,7 @@ static void generate_json(const std::string &output) {
     if (detected_status.size() == 0) {
         json.push_back("]\n}");
     } else {
-        for (u8 i = 0; i < detected_status.size(); i++) {
+        for (size_t i = 0; i < detected_status.size(); i++) {
             json.push_back("\n\t\t\"");
             json.push_back(VM::flag_to_string(detected_status.at(i)));
